float => double

+++ svg_attributes.cpp
@@ -44,7 +44,7 @@ Attributes::Attributes( Object* object )
 
 ///////////////////////////////////////////////////////////////////////////////
 
-Attributes* Attributes::SetOpacity( float opacity )
+Attributes* Attributes::SetOpacity( double opacity )
 {
   if ( opacity < 0.0 ) opacity = 0.0;
   if ( opacity > 1.0 ) opacity = 1.0;
